# By Thomas Kirsten (20) and others
# Via KA101 (11) and others
* upstream/master: (55 commits)
  Added the cvd machine to trader avoid
  Remove useless warning supressors that have no effect on c++ code.
  Style change I forgot to add to previous merge.
  Spelling fix in modified mininuke description
  Allows NPCs to spend cash
  De-slur NPCs
  Fix Wsign-compare in item.cpp and item_group.cpp
  Make clang++ happy by avoiding some ambiguity
  fix Wsign-compare in melee.cpp
  Fix Wswitch in colors.cpp
  fix Wsign-compare in dependency_tree.cpp
  fix of Wsign-compare in uistate.h and bionics.cpp
  Remove duplicate behavior.
  Remove default_size_tile_atlas.
  fix of Wsign. computer.cpp cleaned.
  Fix misc warnings in artifact.cpp
  fix Wsign warnings. npc.h cleaned
  fix Wsign warnings. input.h clean now.
  fix Wswitch in computer
  fix incorrect refreshing of pick-up caption
  ...

Conflicts:
	Makefile
	data/json/bionics.json
	data/json/professions.json
	src/construction.cpp
	src/crafting.cpp
	src/melee.cpp
	src/npctalk.cpp
	src/veh_interact.cpp
	src/vehicle.cpp

+++ src/construction.cpp
@@ -246,7 +246,7 @@ void construction_menu()
                 }
                 // Print components
                 posx = 33;
-                for (int i = 0; i < current_con->components.size(); i++) {
+                for (size_t i = 0; i < current_con->components.size(); i++) {
                     total_inv.has_any_components(current_con->components[i]);
                     mvwprintz(w_con, posy, posx - 2, c_white, ">");
                     for (unsigned j = 0; j < current_con->components[i].size(); j++) {
